split create.cs into a floor and a util class
